Seperate a few commands into specific files
