+++ freqtrade/commands/__init__.py
@@ -1,6 +1,10 @@
-from .utils import (setup_utils_configuration, start_create_userdir,  # noqa: F401
-                    start_download_data, start_hyperopt_list,
-                    start_hyperopt_show, start_list_exchanges,
-                    start_list_markets, start_list_strategies,
-                    start_list_timeframes, start_new_hyperopt,
-                    start_new_strategy, start_test_pairlist, start_trading)
+from .hyperopt_commands import (start_hyperopt_list, start_hyperopt_show)  # noqa: 401
+
+from .list_commands import (start_list_exchanges,  # noqa: F401
+                            start_list_markets, start_list_strategies,
+                            start_list_timeframes)
+from .utils import setup_utils_configuration  # noqa: F401
+from .utils import (start_download_data,  # noqa: F401
+                    start_test_pairlist)
+from .deploy_commands import (start_new_hyperopt, start_new_strategy, start_create_userdir)  # noqa: F401
+from .trade_commands import start_trading  # noqa: F401
